app/preprocessing/assistants.py
import autogen
from typing import Dict, Any

class PreprocessingAssistants:
    """Class for preprocessing assistants"""
    
    def __init__(self, config_list: Dict[str, Any]):
        """
        Initialize preprocessing assistants
        
        Args:
            config_list: LLM configuration dictionary
        """
        self.config = {"config_list": [config_list]}  # Wrap in config_list structure expected by autogen
        self.connection_parsing_assistant = self._create_connection_parsing_assistant()
        self.schema_parsing_assistant = self._create_schema_parsing_assistant()
        self.table_parsing_assistant = self._create_table_parsing_assistant()

    def _create_connection_parsing_assistant(self):
        return autogen.AssistantAgent(
            name="Parsing_agent",
            human_input_mode="NEVER",
            max_consecutive_auto_reply=3,
            llm_config=self.config,
            system_message=self._get_connection_parsing_prompt(),
            code_execution_config=False
        )

    def _create_schema_parsing_assistant(self):
        return autogen.AssistantAgent(
            name="Schema_Parsing_agent",
            llm_config=self.config,
            system_message=self._get_schema_parsing_prompt(),
            max_consecutive_auto_reply=4,
            code_execution_config=False
        )

    def _create_table_parsing_assistant(self):
        return autogen.AssistantAgent(
            name="Table_Parsing_agent",
            llm_config=self.config,
            system_message=self._get_table_parsing_prompt(),
            max_consecutive_auto_reply=4,
            code_execution_config=False
        )

    @staticmethod
    def _get_connection_parsing_prompt() -> str:
        return """You're an expert at identifying the connection name from the text...
        # ... rest of the connection parsing prompt ..."""

    @staticmethod
    def _get_schema_parsing_prompt() -> str:
        return """You help the human identify the inferred SCHEMA name...
        # ... rest of the schema parsing prompt ..."""

    @staticmethod
    def _get_table_parsing_prompt() -> str:
        return """You help the human identify the inferred TABLE name...
        # ... rest of the table parsing prompt ..."""

    async def process(self, message: str) -> str:
        """Process a message using the assistants"""
        # TODO: Implement actual preprocessing
        return message 