requirements.txt
tabulate==0.9.0
python-dotenv==1.0.1
#groq==0.11.0
#cerebras_cloud_sdk==1.10.0
typing-inspect==0.9.0
typing_extensions==4.12.2
duckdb==1.1.3
gradio==5.5.0
gradio_client==1.4.2
openai-gradio==0.0.3
google-generativeai==0.8.3
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.71.1
openai==1.30.1
fastapi<1
aiohttp==3.10.10
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
opentelemetry-util-http==0.49b0
safehttpx==0.1.1
uvicorn>=0.18.3
autogen-agentchat==0.2.39
